[Spy] Extract TypesControl from the MainWindow
