+++ RemoteNetSpy/MainWindow.xaml.cs
@@ -325,10 +325,9 @@ namespace RemoteNetSpy
                 }
             }
 
-
-
             // Reapply filter for types
-            filterBox_TextChanged(typesFilterBox, null);
+            // TODO:
+            //filterBox_TextChanged(typesFilterBox, null);
         }
 
         private async Task<List<DumpedType>> GetTypesList()
@@ -357,45 +356,45 @@ namespace RemoteNetSpy
             return tempList;
         }
 
-        private async void typesListBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            _currSelectedType = (_currSelectedType);
-            string type = _currSelectedType?.FullTypeName;
-            if (type == null)
-            {
-                membersListBox.ItemsSource = null;
-                return;
-            }
+        //private async void typesListBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        //{
+        //    _currSelectedType = (_currSelectedType);
+        //    string type = _currSelectedType?.FullTypeName;
+        //    if (type == null)
+        //    {
+        //        membersListBox.ItemsSource = null;
+        //        return;
+        //    }
 
 
-            var x = CliWrap.Cli.Wrap("rnet-dump.exe")
-                .WithArguments($"members -t {TargetPid} -q \"{type}\" -n true " + UnmanagedFlagIfNeeded())
-                .WithValidation(CommandResultValidation.None)
-                .ExecuteBufferedAsync();
-            var res = await x.Task;
-            var xx = res.StandardOutput.Split('\n', StringSplitOptions.RemoveEmptyEntries)
-                .SkipWhile(line => !line.Contains("Members of "))
-                .Skip(1)
-                .Select(str => str.Trim());
+        //    var x = CliWrap.Cli.Wrap("rnet-dump.exe")
+        //        .WithArguments($"members -t {TargetPid} -q \"{type}\" -n true " + UnmanagedFlagIfNeeded())
+        //        .WithValidation(CommandResultValidation.None)
+        //        .ExecuteBufferedAsync();
+        //    var res = await x.Task;
+        //    var xx = res.StandardOutput.Split('\n', StringSplitOptions.RemoveEmptyEntries)
+        //        .SkipWhile(line => !line.Contains("Members of "))
+        //        .Skip(1)
+        //        .Select(str => str.Trim());
 
-            List<string> rawLinesList = xx.ToList();
-            List<DumpedMember> dumpedMembers = new List<DumpedMember>();
-            for (int i = 0; i < rawLinesList.Count; i += 2)
-            {
-                DumpedMember dumpedMember = new DumpedMember()
-                {
-                    RawName = rawLinesList[i],
-                    NormalizedName = rawLinesList[i + 1]
-                };
-                dumpedMembers.Add(dumpedMember);
-            }
+        //    List<string> rawLinesList = xx.ToList();
+        //    List<DumpedMember> dumpedMembers = new List<DumpedMember>();
+        //    for (int i = 0; i < rawLinesList.Count; i += 2)
+        //    {
+        //        DumpedMember dumpedMember = new DumpedMember()
+        //        {
+        //            RawName = rawLinesList[i],
+        //            NormalizedName = rawLinesList[i + 1]
+        //        };
+        //        dumpedMembers.Add(dumpedMember);
+        //    }
 
-            dumpedMembers.Sort(CompareDumperMembers);
+        //    dumpedMembers.Sort(CompareDumperMembers);
 
-            membersListBox.ItemsSource = dumpedMembers;
+        //    membersListBox.ItemsSource = dumpedMembers;
 
-            filterBox_TextChanged(membersFilterBox, null);
-        }
+        //    filterBox_TextChanged(membersFilterBox, null);
+        //}
 
         private int CompareDumperMembers(DumpedMember member1, DumpedMember member2)
         {
@@ -510,30 +509,30 @@ namespace RemoteNetSpy
             Regex r = null;
 
             ListBox associatedBox = null;
-            if (sender == typesFilterBox)
-            {
-                associatedBox = typesListBox;
-                matchCase = _matchCaseTypes;
-                useRegex = _regexTypes;
-                onlyTypesInHeap = _onlyTypesInHeap;
+            //if (sender == typesFilterBox)
+            //{
+            //    associatedBox = typesListBox;
+            //    matchCase = _matchCaseTypes;
+            //    useRegex = _regexTypes;
+            //    onlyTypesInHeap = _onlyTypesInHeap;
 
-                if (useRegex)
-                {
-                    try
-                    {
-                        string tempFilter = (sender as TextBox)?.Text;
-                        r = new Regex(tempFilter);
-                    }
-                    catch
-                    {
-                        typesFilterBoxBorder.BorderBrush = new SolidColorBrush(Color.FromRgb(255, 153, 164));
-                        return;
-                    }
-                }
+            //    if (useRegex)
+            //    {
+            //        try
+            //        {
+            //            string tempFilter = (sender as TextBox)?.Text;
+            //            r = new Regex(tempFilter);
+            //        }
+            //        catch
+            //        {
+            //            typesFilterBoxBorder.BorderBrush = new SolidColorBrush(Color.FromRgb(255, 153, 164));
+            //            return;
+            //        }
+            //    }
 
-                // No errors in the types filter, reset border
-                typesFilterBoxBorder.BorderBrush = null;
-            }
+            //    // No errors in the types filter, reset border
+            //    typesFilterBoxBorder.BorderBrush = null;
+            //}
 
             if (sender == assembliesFilterBox)
             {
@@ -571,15 +570,15 @@ namespace RemoteNetSpy
                         return (o as DumpedMember)?.NormalizedName?.Contains(filter, comp) == true;
                     }
 
-                    if (sender == typesFilterBox)
-                    {
-                        string input = _dumpedTypeToDescription.Convert(o, null, null, null) as string;
-                        if (onlyTypesInHeap && !HeapInstancesRegex().IsMatch(input))
-                            return false;
-                        if (!useRegex)
-                            return input?.Contains(filter, comp) == true;
-                        return r.IsMatch(input);
-                    }
+                    //if (sender == typesFilterBox)
+                    //{
+                    //    string input = _dumpedTypeToDescription.Convert(o, null, null, null) as string;
+                    //    if (onlyTypesInHeap && !HeapInstancesRegex().IsMatch(input))
+                    //        return false;
+                    //    if (!useRegex)
+                    //        return input?.Contains(filter, comp) == true;
+                    //    return r.IsMatch(input);
+                    //}
 
                     if (sender == assembliesFilterBox)
                     {
@@ -593,8 +592,6 @@ namespace RemoteNetSpy
 
         private void clearTypesFilterButton_OnClick(object sender, RoutedEventArgs e)
         {
-            if (sender == clearTypesFilterButton)
-                typesFilterBox.Clear();
             if (sender == clearAssembliesFilterButton)
                 assembliesFilterBox.Clear();
             if (sender == clearMembersFilterButton)
@@ -604,7 +601,8 @@ namespace RemoteNetSpy
         private async void ProcsRefreshButton_OnClick(object sender, RoutedEventArgs e)
         {
             membersListBox.ItemsSource = null;
-            typesListBox.ItemsSource = null;
+            _typesModel.SelectedType = null;
+            _typesModel.Types = null;
             assembliesListBox.ItemsSource = null;
             heapInstancesListBox.ItemsSource = null;
             _traceList.Clear();
@@ -870,7 +868,7 @@ namespace RemoteNetSpy
                     typesAndInstancesCount[heapObjectType]++;
             }
 
-            string lastSelected = (typesListBox?.SelectedItem as DumpedType)?.FullTypeName;
+            string lastSelected = _currSelectedType?.FullTypeName;
             DumpedType typeToReselect = null;
 
             List<DumpedType> dumpedTypes = new List<DumpedType>();
@@ -897,14 +895,15 @@ namespace RemoteNetSpy
                 }
             }
 
-            typesListBox.ItemsSource = dumpedTypes;
+            _typesModel.Types = new ObservableCollection<DumpedType>(dumpedTypes);
             if (typeToReselect != null)
             {
-                typesListBox.SelectedItem = typeToReselect;
+                _typesModel.SelectedType = typeToReselect;
             }
 
             // Reapply filter
-            filterBox_TextChanged(typesFilterBox, null);
+            //TODO:
+            //filterBox_TextChanged(typesFilterBox, null);
 
             spinner1.Visibility = Visibility.Collapsed;
             countLabel.Foreground = originalBrush;
@@ -953,27 +952,7 @@ namespace RemoteNetSpy
             Clipboard.SetText(typeName);
         }
 
-        private void TypeMenuItem_OnClick(object sender, RoutedEventArgs e)
-        {
-            MenuItem mi = sender as MenuItem;
-            string typeName = (mi.DataContext as DumpedType).FullTypeName;
-            Clipboard.SetText(typeName);
-        }
 
-        private void GoToAssemblyMenuItem_OnClick(object sender, RoutedEventArgs e)
-        {
-            MenuItem mi = sender as MenuItem;
-            string assembly = (mi.DataContext as DumpedType).Assembly;
-            AssemblyModel matchingAssembly = (assembliesListBox.ItemsSource as List<AssemblyModel>).FirstOrDefault(x => x.Name == assembly);
-            int index = assembliesListBox.Items.IndexOf(matchingAssembly);
-
-            // Trick to scroll to our selected item from the BOTTOM
-            double singleListItemHeight = assembliesListBox.FindVisualChildren<ListBoxItem>().First().ActualHeight;
-            double numItemsShown = assembliesListBox.ActualHeight / singleListItemHeight;
-            var furtherDownItem = assembliesListBox.Items[Math.Min(index + (int)numItemsShown - 2, assembliesListBox.Items.Count - 1)];
-            assembliesListBox.SelectedItem = matchingAssembly;
-            assembliesListBox.ScrollIntoView(furtherDownItem);
-        }
 
         private void MemberMenuItem_OnClick(object sender, RoutedEventArgs e)
         {
@@ -1268,7 +1247,6 @@ dynamic dro = ro.Dynamify();
                 var selectedType = (sender as TypesModel).SelectedType;
                 if (selectedType != null)
                 {
-                    _currSelectedType = selectedType;
                     string type = _currSelectedType?.FullTypeName;
                     if (type == null)
                     {
@@ -1306,6 +1284,21 @@ dynamic dro = ro.Dynamify();
                 }
             }
         }
+
+#pragma warning disable IDE0051 // Remove unused private members
+        private void TypesControl_GoToAssemblyInvoked(string assembly)
+        {
+            AssemblyModel matchingAssembly = (assembliesListBox.ItemsSource as List<AssemblyModel>).FirstOrDefault(x => x.Name == assembly);
+            int index = assembliesListBox.Items.IndexOf(matchingAssembly);
+
+            //Trick to scroll to our selected item from the BOTTOM
+            double singleListItemHeight = assembliesListBox.FindVisualChildren<ListBoxItem>().First().ActualHeight;
+            double numItemsShown = assembliesListBox.ActualHeight / singleListItemHeight;
+            var furtherDownItem = assembliesListBox.Items[Math.Min(index + (int)numItemsShown - 2, assembliesListBox.Items.Count - 1)];
+            assembliesListBox.SelectedItem = matchingAssembly;
+            assembliesListBox.ScrollIntoView(furtherDownItem);
+        }
+#pragma warning restore IDE0051 // Remove unused private members
     }
 
     public static class VisualTreeHelperExtensions
